+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- View the schema of a DuckDB database in the sidebar.
+- Run queries and view the results.
+
 ## [0.0.1] - 2023-05-02
 
 -   Use the DuckDB CLI